extract slp from legacy.cpp, upgrade nominmax to standard header
nominmax is used so often that i prefer to use <> for including the
file. also, default exception handling is now provided by core/except.hpp
