+++ game/src/core/except.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+/* lazy default error handling headers */
+
+#include <stdexcept>
+#include <string>
