+++ game/src/legacy.hpp
@@ -8,7 +8,7 @@
 #include <memory>
 #include <map>
 
-#include "nominmax.hpp"
+#include <minmax.hpp>
 
 #include <SDL2/SDL_pixels.h>
 #include <SDL2/SDL_surface.h>
