+++ game/src/ui/menubar.cpp
@@ -4,7 +4,7 @@
 #include <imgui_impl_sdl.h>
 #include <imgui_impl_opengl3.h>
 
-#include "../nominmax.hpp"
+#include <minmax.hpp>
 
 namespace aoe {
